Implement wait_n async routine for spawning wait_random multiple times
This commit adds a new async routine called wait_n to the 1-concurrent_coroutines module. The wait_n routine takes in two integer arguments, n and max_delay, and spawns the wait_random coroutine n times with the specified max_delay. It returns a list of all the delays in ascending order without using sort() due to concurrency.

The purpose of this change is to provide a convenient way to execute multiple instances of wait_random concurrently and retrieve the delays in a sorted order.